refactor(milvus): rename mutation helper module to utils
